refactoring and added margin variant to textfield
