+++ src/index.ts
@@ -24,7 +24,7 @@ export * from './components/ServiceTiles';
 export * from './components/StatusChip';
 export * from './components/Switch';
 export * from './components/TabNavigation';
-export * from './components/TextEditor';
+export * from './components/TextEditor/TextEditor';
 export * from './components/TextField';
 export * from './components/LogTable';
 export * from './components/Checkbox';
